Epoch conversion: civilFromDaysSince0000 in int
The day count of ten thousand years is 22 bits, so the seven divisions
of the era/year/month arithmetic can be 32-bit. The instruction count
did not move (a 32-bit magic division is a movslq, imul and sar where
the 64-bit one is a movabs, imul and sar), nor did the time; kept with
S7.1 as the simpler code. The conversion is a dependency chain and
that, not the count, is what its time is made of - see the S7 findings.

perf-log: S7.2

+++ src/main/java/com/ethlo/time/internal/fixed/ITUEpochParser.java
@@ -191,7 +191,7 @@ public class ITUEpochParser
     private static DateTime toDateTime(final long secondsSince0000, final int nano, final Field field, final int fractionDigits, final int parseLength)
     {
         final long days = secondsSince0000 / 86_400;
-        final int date = DateTimeMath.civilFromDaysSince0000(days);
+        final int date = DateTimeMath.civilFromDaysSince0000((int) days);
         final int secondOfDay = (int) (secondsSince0000 - days * 86_400);
         return new DateTime(field, DateTimeMath.packedYear(date), DateTimeMath.packedMonth(date), DateTimeMath.packedDay(date), secondOfDay / 3_600, (secondOfDay / 60) % 60, secondOfDay % 60, nano, TimezoneOffset.UTC, fractionDigits, parseLength);
     }
@@ -199,7 +199,7 @@ public class ITUEpochParser
     private static void fill(final MutableDateTimeBuffer out, final long secondsSince0000, final int nano, final Field field, final int fractionDigits, final int parseLength)
     {
         final long days = secondsSince0000 / 86_400;
-        final int date = DateTimeMath.civilFromDaysSince0000(days);
+        final int date = DateTimeMath.civilFromDaysSince0000((int) days);
         final int secondOfDay = (int) (secondsSince0000 - days * 86_400);
         out.set(field, DateTimeMath.packedYear(date), DateTimeMath.packedMonth(date), DateTimeMath.packedDay(date), secondOfDay / 3_600, (secondOfDay / 60) % 60, secondOfDay % 60, nano, fractionDigits, 0, parseLength);
     }
